requirements.txt
Flask==3.0.3
flask-cors==4.0.0
openai-whisper==20231117